Add fonts modules
Includes a modules for fonts that can be installed as a distribution's
package and another modules for TrueType fonts that can be downloaded
from the Internet.

The latter installs the fonts only for the current user. This might
change in the future, as I am not fully convinced this is what I want.
To make this change easier, I have extracted the location to a different
function, so I can change it in one place only.