Use public dotfiles URL to avoid password prompt
When cloning my dotfiles, I normally have not gone through the process
of setting ssh keys and add them to github, so it would ask me for a
password. I am switching to a public URL instead, which means I'll have
to add the private one later if I want to push anything, but that's
Okay.