Don't overwrite output with state indicators
Previously, the end of module would go up one line and print the status
at the end. This worked well for noop modules, but for modules that are
applied and generate output, this was overwriting that output.

This changeset reprints the module name if the module has been applied
and then the corresponding status, making sure the commands output is
never overwritten.

Fixes #9